+++ tests/integration/test_monte_carlo_thevenin.py
@@ -58,12 +58,12 @@ class TestSamplingThevenin:
         return pybop.Parameters(
             pybop.Parameter(
                 "R0 [Ohm]",
-                prior=pybop.Uniform(1e-2, 8e-2),
+                prior=pybop.Uniform(1e-3, 9e-2),
                 bounds=[1e-4, 1e-1],
             ),
             pybop.Parameter(
                 "R1 [Ohm]",
-                prior=pybop.Uniform(1e-2, 8e-2),
+                prior=pybop.Uniform(1e-3, 9e-2),
                 bounds=[1e-4, 1e-1],
             ),
         )